+++ src/emu/cpu/alph8201/8201dasm.c
@@ -321,7 +321,7 @@ static void InitDasm8201(void)
 		Op[i].type  = type;
 
 		/* 2 byte code ? */
-		while (isspace(*p)) p++;
+		while (isspace((UINT8)*p)) p++;
 		if( (*p) )
 			Op[i].type |= 0x10;
 		/* number of param */
