+++ src/lib/util/corefile.c
@@ -884,7 +884,7 @@ int core_filename_ends_with(const char *filename, const char *extension)
 
 	/* work backwards checking for a match */
 	while (extlen > 0)
-		if (tolower(filename[--namelen]) != tolower(extension[--extlen]))
+		if (tolower((UINT8)filename[--namelen]) != tolower((UINT8)extension[--extlen]))
 		{
 			matches = FALSE;
 			break;
